Update Welcome to Java!
Welcome to the world of Java! In this challenge, we practice printing to stdout.
The code stubs in your editor declare a Solution class and a main method. Complete the main method by copying the two lines of code below and pasting them inside the body of your main method.

Output Format

You must print two lines of output:

Print Hello, World. on the first line.
Print Hello, Java. on the second line.